Remove posts submodule
Initial reasoning was that it would be nice to view the history of just the
changes to posts.  However, github has this feature!

The only other benefit is that you could check out the site without pulling the
posts. This is not beneficial enough to put up with the pain of git submodules