Load bold and italic fonts, use single request
By default the Google Fonts API sends only a stylesheet that includes
the regular weight and style. If you want bold and italic, you have to
explicitly ask for them. Also, all of the requests can be collapsed into
one to reduce the number of roundtrips (when not using http/2) and to
reduce the page size.